Storage Prototype 2
Code commented out is designed to find the profile of the user, but it can't search by the email and the object id is not set to the users id when the profile is created. Thinking about sending through the information of the profiles id as a property in the user class with user.setProperty().